backend/requirements.txt
alembic==1.7.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
clickclick==20.10.2
colorama==0.4.4
connexion==2.7.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
idna==2.10
inflection==0.5.1
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow==3.12.1
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
psycopg2-binary==2.9.1
pyrsistent==0.18.0
python-dotenv==0.19.1
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.23
stripe==2.61.0
swagger-ui-bundle==0.0.8
urllib3==1.26.6
Werkzeug==2.0.1
